Moved composition logic to `EssentialApp` (composition root), was in `EssentialFeediOS`
