+++ EssentialApp/EssentialApp/WeakRefVirtualProxy.swift
